refactor: consolidate all *Utils into ui/util/
domain/ had 9 UI helper files (Activity/Fragment/View extensions)
alongside 3 true domain files. Move all *Utils to ui/util/ so domain/
contains only use cases, models, and pure business logic.

domain/ after: AppStartUtils, Icon, ObserveIconListUseCase
ui/util/ gains: AutoClearedUtils, DrawerUtils, LifecycleUtils,
  OnboardingUtils, SearchHighlighter, SearchUtils, SnackBarUtils,
  SplashUtils, ToastUtils, TransitionUtils, URLUtils
