+++ app/src/main/java/de/lemke/oneuisample/ui/OOBEActivity.kt
@@ -24,6 +24,8 @@ import de.lemke.oneuisample.BuildConfig
 import de.lemke.oneuisample.R
 import de.lemke.oneuisample.data.UserSettingsRepository
 import de.lemke.oneuisample.databinding.ActivityOobeBinding
+import de.lemke.oneuisample.ui.util.EXTRA_VERSION_CODE
+import de.lemke.oneuisample.ui.util.EXTRA_VERSION_NAME
 import dev.oneuiproject.oneui.widget.OnboardingTipsItemView
 import javax.inject.Inject
 import kotlin.time.Duration.Companion.milliseconds
