+++ app/src/main/java/de/lemke/oneuisample/ui/fragments/SubtabWidgets.kt
@@ -21,9 +21,9 @@ import dagger.hilt.android.AndroidEntryPoint
 import de.lemke.oneuisample.R
 import de.lemke.oneuisample.databinding.FragmentTabDesignSubtabWidgetsBinding
 import de.lemke.oneuisample.databinding.FragmentTabDesignSubtabWidgetsBinding.inflate
-import de.lemke.oneuisample.domain.autoCleared
-import de.lemke.oneuisample.domain.suggestiveSnackBar
 import de.lemke.oneuisample.ui.MainActivity
+import de.lemke.oneuisample.ui.util.autoCleared
+import de.lemke.oneuisample.ui.util.suggestiveSnackBar
 import dev.oneuiproject.oneui.ktx.setEntries
 
 @AndroidEntryPoint
