+++ app/src/main/java/de/lemke/oneuisample/ui/util/DrawerUtils.kt
@@ -15,7 +15,7 @@
  */
 @file:Suppress("unused")
 
-package de.lemke.oneuisample.ui
+package de.lemke.oneuisample.ui.util
 
 import android.content.Intent
 import android.os.Bundle
