+++ app/src/main/java/de/lemke/oneuisample/ui/util/OnboardingUtils.kt
@@ -1,12 +1,11 @@
-package de.lemke.oneuisample.ui
+package de.lemke.oneuisample.ui.util
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import de.lemke.oneuisample.data.UserSettingsRepository
 import de.lemke.oneuisample.domain.AppStart
 import de.lemke.oneuisample.domain.checkAppStart
-import de.lemke.oneuisample.domain.finishWithFade
-import de.lemke.oneuisample.domain.overrideFadeOpenTransition
+import de.lemke.oneuisample.ui.OOBEActivity
 
 const val EXTRA_SKIP_ONBOARDING = "skipOnboarding"
 const val EXTRA_VERSION_CODE = "versionCode"
