+++ app/src/main/java/de/lemke/oneuisample/ui/util/SnackBarUtils.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package de.lemke.oneuisample.domain
+package de.lemke.oneuisample.ui.util
 
 import android.R.id.content
 import android.app.Activity
