+++ app/src/main/java/de/lemke/oneuisample/ui/util/TransitionUtils.kt
@@ -1,4 +1,4 @@
-package de.lemke.oneuisample.domain
+package de.lemke.oneuisample.ui.util
 
 import android.R.anim.fade_in
 import android.R.anim.fade_out
