+++ app/src/main/java/de/lemke/oneuisample/ui/util/URLUtils.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package de.lemke.oneuisample.domain
+package de.lemke.oneuisample.ui.util
 
 import android.content.ActivityNotFoundException
 import android.content.Context
